requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Pillow==2.6.1
PyYAML==3.11
argparse==1.2.1
django-autoslug==1.7.2
django-composition==0.2
django-generic-images==0.36.1
django-grappelli==2.6.3
django-mptt==0.6.1
django-storages==1.1.8
django-tinymce==1.5.3
django-transaction-hooks==0.2
django-treebeard==2.0
eav-django==1.4.7
wsgiref==0.1.2
